Changed doubles to floats to comply with taproot better
All subsystems except for indexer are now float-based

+++ MCB-project/src/subsystems/drivetrain/DrivetrainSubsystem.cpp
@@ -66,14 +66,14 @@ void DrivetrainSubsystem::refresh()
     drivers->djiMotorTxHandler
         .encodeAndSendCanData();  // Processes these motor speed changes into CAN signal
 }
-static double motorOneSpeed, motorTwoSpeed, motorThreeSpeed, motorFourSpeed = 0;
+static float motorOneSpeed, motorTwoSpeed, motorThreeSpeed, motorFourSpeed = 0;
 
 void DrivetrainSubsystem::moveDriveTrain(
-    double turnSpeed,
-    double translationSpeed,
-    double translationAngle)
+    float turnSpeed,
+    float translationSpeed,
+    float translationAngle)
 {
-    double angularOffset = (motor_one.getShaftRPM() + motor_three.getShaftRPM()) / 2 / 14000.0;
+    float angularOffset = (motor_one.getShaftRPM() + motor_three.getShaftRPM()) / 2 / 14000.0;
 
     convertTranslationSpeedToMotorSpeeds(
         translationSpeed,
@@ -88,45 +88,45 @@ void DrivetrainSubsystem::setMotorSpeeds()
 
     powerLimit += limitIncrease;  // add limit increase
 
-    double w1 = motorOneRPM * (2 * M_PI / 60);    // Convert RPM to rad/s
-    double w2 = motorTwoRPM * (2 * M_PI / 60);    // Convert RPM to rad/s
-    double w3 = motorThreeRPM * (2 * M_PI / 60);  // Convert RPM to rad/s
-    double w4 = motorFourRPM * (2 * M_PI / 60);   // Convert RPM to rad/s
+    float w1 = motorOneRPM * (2 * M_PI / 60);    // Convert RPM to rad/s
+    float w2 = motorTwoRPM * (2 * M_PI / 60);    // Convert RPM to rad/s
+    float w3 = motorThreeRPM * (2 * M_PI / 60);  // Convert RPM to rad/s
+    float w4 = motorFourRPM * (2 * M_PI / 60);   // Convert RPM to rad/s
 
     // set the outputs according to the pid controller and get currents adjusted for saturation
     pidController.runControllerDerivateError(motorOneSpeed - motorOneRPM, 1);
-    double I1t = std::clamp(
-        static_cast<double>(pidController.getOutput()) / 819.2,
+    float I1t = std::clamp(
+        static_cast<float>(pidController.getOutput()) / 819.2f,
         -abs(VOLT_MAX - KB * w1) / RA,
         abs(VOLT_MAX - KB * w1) / RA);
 
     pidController.runControllerDerivateError(motorTwoSpeed - motorTwoRPM, 1);
-    double I2t = std::clamp(
-        static_cast<double>(pidController.getOutput()) / 819.2,
+    float I2t = std::clamp(
+        static_cast<float>(pidController.getOutput()) / 819.2f,
         -abs(VOLT_MAX - KB * w2) / RA,
         abs(VOLT_MAX - KB * w2) / RA);
 
     pidController.runControllerDerivateError(motorThreeSpeed - motorThreeRPM, 1);
-    double I3t = std::clamp(
-        static_cast<double>(pidController.getOutput()) / 819.2,
+    float I3t = std::clamp(
+        static_cast<float>(pidController.getOutput()) / 819.2f,
         -abs(VOLT_MAX - KB * w3) / RA,
         abs(VOLT_MAX - KB * w3) / RA);
 
     pidController.runControllerDerivateError(motorFourSpeed - motorFourRPM, 1);
-    double I4t = std::clamp(
-        static_cast<double>(pidController.getOutput()) / 819.2,
+    float I4t = std::clamp(
+        static_cast<float>(pidController.getOutput()) / 819.2f,
         -abs(VOLT_MAX - KB * w4) / RA,
         abs(VOLT_MAX - KB * w4) / RA);
 
     // Calculate total power requested
-    double totalPowerRequested = 0.7 * RA * (I1t * I1t + I2t * I2t + I3t * I3t + I4t * I4t) +
+    float totalPowerRequested = 0.7 * RA * (I1t * I1t + I2t * I2t + I3t * I3t + I4t * I4t) +
                                  VELO_LOSS * (abs(w1) + abs(w2) + abs(w3) + abs(w4));
-    // double totalPowerRequested =
+    // float totalPowerRequested =
     //   RA * (I1t * I1t + I2t * I2t + I3t * I3t + I4t * I4t) + KT * (abs(w1 * I1t) + abs(w2 * I2t)
     //   + abs(w3 * I3t) + abs(w4 * I4t));
 
     // Scale currents if power limit is exceeded
-    double scale = std::max((double)1.0, (totalPowerRequested + IDLE_DRAW) / powerLimit);
+    float scale = std::max((float)1.0, (totalPowerRequested + IDLE_DRAW) / powerLimit);
 
     motor_one.setDesiredOutput(static_cast<int32_t>(I1t * 819.2 / scale));
     motor_two.setDesiredOutput(static_cast<int32_t>(I2t * 819.2 / scale));
@@ -144,8 +144,8 @@ void DrivetrainSubsystem::stopMotors()
 }
 
 void DrivetrainSubsystem::convertTranslationSpeedToMotorSpeeds(
-    double translationSpeed,
-    double translationAngle)
+    float translationSpeed,
+    float translationAngle)
 {
     motorOneSpeed = translationSpeed * sin(translationAngle + (PI / 4));
     motorTwoSpeed = translationSpeed * sin(translationAngle - (PI / 4));
@@ -153,7 +153,7 @@ void DrivetrainSubsystem::convertTranslationSpeedToMotorSpeeds(
     motorFourSpeed = translationSpeed * sin(translationAngle + (PI / 4));
 }
 
-void DrivetrainSubsystem::adjustMotorSpeedWithTurnSpeed(double turnSpeed)
+void DrivetrainSubsystem::adjustMotorSpeedWithTurnSpeed(float turnSpeed)
 {
     motorOneSpeed += turnSpeed;
     motorTwoSpeed -= turnSpeed;
